Add PointOfInterestForCreationDto Class
Added the for creation dto to handle post of new point of interest

Add CreatePointOfInterest Method

Added method and all neccessary logic to to PointsOfInterestController to handle the creation of a point of interest to a city.

Added Parameter to HttpGet Decoration

Added a name paramater to the HttpGet decoration to the GetPointOfInterest method for reference by the newly-created CreatedAtRoute method in the the CreatePointOfInterest method.